test ta entehay edit course

+++ modules/studiobarg/Category/Http/Controllers/CategoryController.php
@@ -3,13 +3,16 @@
 namespace Studiobarg\Category\Http\Controllers;
 
 use App\Http\Controllers\Controller;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use studiobarg\Category\Http\Requests\CategoryRequest;
+use studiobarg\Category\Models\Category;
 use studiobarg\Category\Repository\categoryRepo;
 use studiobarg\Category\Responses\AjaxResponse;
 
 class CategoryController extends Controller
 {
+    use AuthorizesRequests;
     public $repo;
 
     public function __construct(categoryRepo $categoryRepo)
@@ -19,18 +22,21 @@ class CategoryController extends Controller
 
     public function index()
     {
+        $this->authorize('manage', category::class);
         $categories = $this->repo->all();
         return view('Category::index', compact('categories'));
-    }//End Method
+    } //End Method
 
     public function store(CategoryRequest $request)
     {
+        $this->authorize('manage', category::class);
         $this->repo->store($request);
         return back();
     }//End Method
 
     public function edit($categoryId)
     {
+        $this->authorize('manage', category::class);
         $category = $this->repo->findById($categoryId);
         $categories = $this->repo->allExceptById($categoryId);
         return view('Category::edit', compact('category', 'categories'));
@@ -38,12 +44,14 @@ class CategoryController extends Controller
 
     public function update($categoryId, Request $request)
     {
+        $this->authorize('manage', category::class);
         $this->repo->update($categoryId, $request);
         return back();
     }//End Method
 
     public function destroy($categoryId)
     {
+        $this->authorize('manage', category::class);
         try {
             $this->repo->delete($categoryId);
             return AjaxResponse::successResponse();
